refactor(dashboard): split territory modal code
